fix(dj): `updateInputSettings({})` should not result in a validation failure
Per the TypeScript definition, both `audio` and `video` fields in the argument object are optional.

The call should simply be a no-op. Today, if `audio` is omitted, audio settings are unchanged, and if `video` is omitted, video settings are unchanged. It would be consistent if omitting both simply resulted in neither settings changing.

Fixing this would allow customers to more cleanly build a camera effects preview view, by letting them do the following without having to special-case the previewer's input settings being empty (which it starts off as):

```
call.updateInputSettings(await previewer.getInputSettings())
```