fix window dependency for custom callbacks to be multi-call instance safe
introduces a new 'customCallbacks' field to the _daily.instances set
for the purpose of passing lambdas/callback functions provided by the
client and needed by the call bundle.

currently, this is only used for the remote and local SdpHooks which
are not documented so possibly, this is really unnecessary. Bumped
the log that occurs when one is provided so we can start tracking usage.