imp(dj): update testCallQuality() to be the bees knees
testCallQuality() now not only uses a secondary call instance to run
the test but generates a test video feed callmachine side, removing
the need for the user to pass in a track. It also automagically stores
the results to be fed in at join time for better initialization of
settings.